- 更新 mybatis-plus 版本 - 调整部分代码 - 添加 GeneratorUtils 代码生成工具类 - 修复 docker-compose 自动化部署数据库中文乱码

+++ api-server/src/main/java/com/mochat/mochat/service/impl/SubsystemServiceImpl.java
@@ -212,7 +212,7 @@ public class SubsystemServiceImpl extends ServiceImpl<SubSystemMapper, UserEntit
         // region 验证手机号唯一
         QueryWrapper<UserEntity> wrapper = new QueryWrapper<>();
         wrapper.eq("phone", parem.getPhone());
-        int cnt = this.baseMapper.selectCount(wrapper);
+        int cnt = Math.toIntExact(this.baseMapper.selectCount(wrapper));
         if (cnt > 0) {
             throw new ParamException("手机号已存在");
         }
@@ -324,7 +324,7 @@ public class SubsystemServiceImpl extends ServiceImpl<SubSystemMapper, UserEntit
         if (!userEntity.getPhone().equals(parem.getPhone())) {
             QueryWrapper<UserEntity> wrapper = new QueryWrapper<>();
             wrapper.eq("phone", parem.getPhone());
-            int cnt = this.baseMapper.selectCount(wrapper);
+            int cnt = Math.toIntExact(this.baseMapper.selectCount(wrapper));
             if (cnt > 0) {
                 throw new ParamException("手机号已存在");
             }
